+++ NEWS.md
@@ -13,6 +13,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * New function to create a ws from datasets and a default specification
 * New function to make a ws crunchable
 * Shiny app to compare WS with button to export table
+* New functions to reverse-engineer a specification
+* New functions to create random specifications
 * New WS example and code to generate it
 
 ### Changed
@@ -21,6 +23,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Rename `cjo` to `td` in functions name (`assign_cjo`, `export_cjo`, `retrieve_cjo`, `import_cjo`, ) and argument (`cjo` -> `td`)
 * The `assign_XXX` and `retrieve_XXX` functions are using `jws` instead of the `ws_path`.
 * `select_regs` takes a new argument `context` to define the different regressors sets to use.
+* `export_XXX` and `import_XXX` functions don't support anymore `ws_name` argument.
 
 ### Fixed